requirements.txt
aws-cdk.core==1.137.0
aws-cdk.aws_iam==1.137.0
aws-cdk.aws_sqs==1.137.0
aws-cdk.aws_sns==1.137.0
aws-cdk.aws_sns_subscriptions==1.137.0
aws-cdk.aws_s3==1.137.0
aws-cdk.assertions==1.137.0
